UpdateProduct

+++ backend/controllers/ProductController.js
@@ -165,7 +165,7 @@ const updateProduct = async(req,res = response)=>{
                     if(req.file.file_ext != 'png' && req.file.file_ext != 'jpg' && req.file.file_ext != 'jpeg' && req.file.file_ext != 'webp'){
                         if(req.file.file_path){fs.unlinkSync(req.file.file_path)}
                     }else{
-                        if(product.banner){
+                        if(product.banner && product.banner ==! undefined && product.banner ==! 'undefined'){
                             fs.unlinkSync('uploads/products/'+product.banner)
                         }
                         data = {
@@ -190,7 +190,6 @@ const updateProduct = async(req,res = response)=>{
 
     } catch (error) {
         if(req.file.file_path){fs.unlinkSync(req.file.file_path)}
-
         res.status(500).json({
             ok: false,
             message: 'Ha ocurrido un error, intenta de nuevo'
